Switch set difference from groovy `-` to Set.removeAll
This difference check verifies if all expected assay values are present
in the database. The original code uses the groovy minus operator, which
uses a quadratic algorithm even if both operands are sets. In datasets
with several thousand assays it can take several seconds to evaluate
this, at least when running in debug mode. The alternative of using
Set.removeAll runs instantly.