AllDataProjection.{data,row}Properties guarantees a stable and sane iteration order
TabSeparatedExporter uses the order from dataProperties for its columns, and probably other exporters/grid view etc do as well. Until now
AllDataProjection just provided a Map for dataProperties and rowProperties without making any iteration guarantees, but the implementation
already used maps with stable iteration in all cases. This commit formalizes that by specifying it in the AllDataProjection and
MultiValueProjection interfaces. Unfortunately there is no widely used interface that signifies predictable iteration order maps.

This commit adds a test for SnpLzAllDataProjection, which is currently the only place where we depend on a certain order.

This replaces the implementation of data/rowProperties for all high dim datatypes by ImmutableMap-s, which are a bit simpler and memory
efficient compared to .asImmutable() views. I have changed all the related core db code to use ImmutableMap instead of Map, to make sure we
keep the stable iteration order. It would be preferable to use a ordered map interface, but there is none that is applicable here. Using a
specific ordered map implementation in core db and noting it in the api docs for core api seems to be the next best thing.